chore: 날짜계산 함수 utils로 분리
